+++ Colour.h
@@ -8,22 +8,22 @@
 
 class Colour {
 private:
-    float red;
-    float green;
-    float blue;
-    float alpha;
+    int red;
+    int green;
+    int blue;
+    int alpha;
 public:
-    Colour(float red_, float green_, float blue_, float alpha_) : red(red_), green(green_), blue(blue_), alpha(alpha_){};
+    Colour(int red_, int green_, int blue_, int alpha_) : red(red_), green(green_), blue(blue_), alpha(alpha_){};
 
-    void SetRed(float red_);
-    void SetGreen(float green_);
-    void SetBlue(float blue_);
-    void SetAlpha(float alpha_);
+    void SetRed(int red_);
+    void SetGreen(int green_);
+    void SetBlue(int blue_);
+    void SetAlpha(int alpha_);
 
-    float GetRed();
-    float GetGreen();
-    float GetBlue();
-    float GetAlpha();
+    int GetRed();
+    int GetGreen();
+    int GetBlue();
+    int GetAlpha();
 
 };
